+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet.
+
+## [v1.20.0] - 2021-08-30
+
 ### Added
 
 - `-v`, `--value` flag to workaround [dash issue](https://github.com/TomWright/dasel/issues/117).
@@ -360,7 +364,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Everything!
 
-[unreleased]: https://github.com/TomWright/dasel/compare/v1.19.0...HEAD
+[unreleased]: https://github.com/TomWright/dasel/compare/v1.20.0...HEAD
+[v1.20.0]: https://github.com/TomWright/dasel/compare/v1.19.0...v1.20.0
 [v1.19.0]: https://github.com/TomWright/dasel/compare/v1.18.0...v1.19.0
 [v1.18.0]: https://github.com/TomWright/dasel/compare/v1.17.0...v1.18.0
 [v1.17.0]: https://github.com/TomWright/dasel/compare/v1.16.1...v1.17.0